fix port values + remove calendar from main (broken)
