+++ Dockerfile
@@ -6,4 +6,4 @@ RUN apk update && apk upgrade && \
 COPY . /app
 RUN ./mvnw package
 CMD ["java", "-jar", "target/spring-0.0.1-SNAPSHOT.jar"]
-EXPOSE 8082
+EXPOSE 8085
