+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
     image: java_springv1
     build: .
     ports:
-      - "8082:8090"
+      - "8085:8085"
     volumes:
        - ./volumes:/volumes
     restart: unless-stopped
