+++ src/main/java/com/nighthawk/spring_portfolio/hacks/calendar/CalendarApiController.java
@@ -1,0 +1,40 @@
+package com.nighthawk.spring_portfolio.hacks.calendar;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+/** Calendar API
+ * Calendar Endpoint: /api/calendar/isLeapYear/2022, Returns: {"year":2020,"isLeapYear":false}
+ */
+@RestController
+@RequestMapping("/api/calendar")
+public class CalendarApiController {
+
+    /** GET isLeapYear endpoint
+     * ObjectMapper throws exceptions on bad JSON
+     *  @throws JsonProcessingException
+     *  @throws JsonMappingException
+     */
+    @GetMapping("/isLeapYear/{year}")
+    public ResponseEntity<JsonNode> getIsLeapYear(@PathVariable int year) throws JsonMappingException, JsonProcessingException {
+      // Backend Year Object
+      Year year_obj = new Year();
+      year_obj.setYear(year);  // evaluates Leap Year
+
+      // Turn Year Object into JSON
+      ObjectMapper mapper = new ObjectMapper(); 
+      JsonNode json = mapper.readTree(year_obj.isLeapYearToString()); // this requires exception handling
+
+      return ResponseEntity.ok(json);  // JSON response, see ExceptionHandlerAdvice for throws
+    }
+
+    // add other methods
+}
