--- src/main/java/com/nighthawk/spring_portfolio/mvc/calendar/CalendarEventRepository.java
@@ -1,9 +1,0 @@
-package com.nighthawk.spring_portfolio.mvc.calendar;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CalendarEventRepository extends JpaRepository<CalendarEvent, Long> {
-    List<CalendarEvent> findAllByOrderByDateAsc();
-}
