+++ src/main/resources/application.properties
@@ -12,7 +12,7 @@ spring.datasource.driver-class-name = org.sqlite.JDBC
 spring.datasource.username = admin
 spring.datasource.password = admin
 
-server.port=8082
+server.port=8085
 spring.session.timeout=30m
 server.servlet.session.cookie.name=sess_java_spring
